refactor(engine): Phase 4 slice 2 — hierarchy.py splits into stages/hierarchy/{graph,ratings,facility_undrawn,unify,enrich}
The 3,363-LOC god module becomes the mandatory stage anatomy: graph.py
(parent/ultimate-parent/facility-graph resolution, cp_lookup_* seals),
ratings.py (dual best-rating inheritance), facility_undrawn.py (synthetic
undrawn rows + MOF waterfall; SA-RW preview stays put until its slice),
unify.py (exposure unification + facility metadata, _OUTPUT_SCHEMA_INVARIANTS
anchor moves with the function), enrich.py (QRRE, rating attach, short-term
override, property coverage, LTV, lending group), resolver.py (verbatim
resolve() recipe + delegating private methods for test back-compat) and
stage.py (the fold adapter). engine/hierarchy.py survives as a 28-line shim
re-exporting HierarchyResolver + _FACILITY_QRRE_COUPLED_COLUMNS.

Function bodies moved verbatim — defensive-surface ratchet metrics are
byte-identical (fill_null 446, presence guards 374, collect_schema 166,
cites 179); max_engine_module_loc banks 3,364 → 2,252. Allowlists re-keyed
(arch_check VALIDATION_ENUM, no-raw-over LINE_ALLOWLIST), citation snapshot
regenerated (2 keys moved to stages.hierarchy.enrich), bytecode-introspection
pin re-pointed at unify.unify_exposures.

Gates: full suite 7,532 passed; parity byte-identical across all four 10k
stress configs; ruff/ty/arch_check green.
