+++ docs/appendix/changelog.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Orchestrator scratch state is gone.** The cross-stage `self._securitisation_resolved` / `self._errors` / `self._ccr_errors` attributes are replaced by typed context artifacts (`SECURITISATION_RESOLVED`, `PIPELINE_ERRORS`, `CCR_ERRORS`, `BRANCH_ERRORS`); the four error channels keep their exact pre-fold merge order and codes (unification is the dedicated error-channel slice, P2.21).
 - **Components are built per run, never cached.** `build_components(config, **overrides)` constructs framework-fresh defaults each run (injected overrides honoured), so the stale-`CRMProcessor` failure mode — a framework switch on a reused orchestrator silently keeping the wrong haircut table, the reason for comparison.py's two-orchestrator workaround — is structurally unrepresentable.
 - **Context-era test surface:** sanctioned builder `tests/fixtures/context.py` (`make_context`), `PipelineContext(` added to the builder-conformance hard lint, fold/registry/context pins in `tests/unit/test_orchestrator_fold.py` + `tests/unit/contracts/test_pipeline_context.py`; the stage-execution tests in `tests/unit/test_pipeline.py` now drive the stage adapters through built contexts.
+- **`engine/hierarchy.py` (3,363 LOC) split into `engine/stages/hierarchy/`** per the mandatory stage anatomy: `graph` (parent/ultimate-parent/facility-graph resolution + the four `cp_lookup_*` seals), `ratings` (dual best-rating inheritance), `facility_undrawn` (synthetic undrawn rows incl. the MOF waterfall; the SA-RW preview stays here until its dedicated slice), `unify` (loans/contingents/facility-undrawn concat + facility metadata), `enrich` (QRRE propagation, rating attach, short-term override, property coverage, LTV, lending group), with a thin `resolver.py` keeping `HierarchyResolver` (verbatim `resolve()` recipe + delegating private methods) and `stage.py` the fold adapter. `engine/hierarchy.py` survives as a 28-line back-compat shim, so the 23+ test files importing `HierarchyResolver` from it are untouched. **Function bodies moved verbatim** — the defensive-surface ratchet metrics are byte-identical (fill_null 446, presence guards 374, collect_schema probes 166) and the parity gate stays byte-identical; `max_engine_module_loc` banks 3,364 → 2,252.
 
 ### Changed (architecture — migration Phase 1: eager stage edges)
 - **Stages now exchange materialised frames; laziness is strictly intra-stage.** `engine/materialise.py` is rewritten around `materialise_edge(lf, config, label)`, called at every stage exit (`hierarchy_exit`, `ccr_exit` when a derivatives book is present, `classifier_exit`, `crm_exit`, `re_split_exit`, and the three calculator branches). The hand-placed barrier inventory — `classifier_output`, `pipeline_pre_branch`, `crm_post_ead_unified`/`_fanout`, `crm_no_guarantee` — is deleted; Two benchmark-justified **intra-stage checkpoints** survive inside CRM: `crm_post_ead` (a controlled A/B showed removing it costs 35–52% on the full-pipeline benchmarks — the collateral lookups re-execute the provisions→CCF→EAD chain without it) and `crm_pre_guarantee_unified` (empirically irreducible on Polars 1.37). Bundle fields stay `LazyFrame`-typed (cheap `.lazy()` wrap) until the Phase 3 producer seal. Byte-identical outputs; the inter-stage plan-depth SIGSEGV class is now unrepresentable. See the rewritten [Stage-Edge Materialisation](../architecture/pipeline-collect-barriers.md) page.
