+++ src/rwa_calc/engine/orchestrator.py
@@ -269,12 +269,12 @@ def build_components(
     from rwa_calc.engine.classifier import ExposureClassifier
     from rwa_calc.engine.crm.processor import CRMProcessor
     from rwa_calc.engine.equity.calculator import EquityCalculator
-    from rwa_calc.engine.hierarchy import HierarchyResolver
     from rwa_calc.engine.irb.calculator import IRBCalculator
     from rwa_calc.engine.re_splitter import RealEstateSplitter
     from rwa_calc.engine.sa.calculator import SACalculator
     from rwa_calc.engine.securitisation.allocator import SecuritisationAllocator
     from rwa_calc.engine.slotting.calculator import SlottingCalculator
+    from rwa_calc.engine.stages.hierarchy import HierarchyResolver
 
     return StageComponents(
         securitisation_allocator=securitisation_allocator or SecuritisationAllocator(),
