+++ src/rwa_calc/engine/stages/hierarchy/enrich.py
@@ -1,0 +1,997 @@
+"""
+Unified-exposure-frame enrichment for the hierarchy stage.
+
+Pipeline position:
+    Loader -> HierarchyResolver (stages/hierarchy) -> Classifier
+    Sub-module of the hierarchy stage package; consumed by ``unify``
+    (QRRE propagation + rating attach) and ``resolver`` (short-term
+    override, property coverage, LTV, lending group).
+
+Key responsibilities:
+- Coalesce QRRE-relevant facility-level columns onto every exposure row
+  (Site B of the ``_FACILITY_QRRE_COUPLED_COLUMNS`` coupling).
+- Attach counterparty rating fields (cqs / pd / internal_pd / model_id).
+- Apply the per-exposure short-term ECAI rating override.
+- Add property-collateral coverage, LTV metadata, and lending-group totals
+  to the unified frame via ``.over()`` window allocation.
+
+References:
+- CRR Art. 131: Short-term rating override for institutional exposures
+- CRR Art. 135 / 136 / 138 / 139 / 140: ECAI rating use and mapping
+- CRR Art. 123(c): retail threshold exclusion (residential property)
+- CRR Art. 126: income-producing CRE risk-weight gate
+- CRR Art. 4(1)(39): group of connected clients (lending group totals)
+- PRA PS1/26 Art. 120(2B) Table 4A / Art. 122(3) Table 6A: short-term tables
+- PRA PS1/26 Art. 121(4): SCRA short-term trade-finance LC window
+- PRA PS1/26 Art. 124(4) / Art. 124A / Art. 124J: non-qualifying RE gate
+"""
+
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING
+
+import polars as pl
+from watchfire import cites
+
+from rwa_calc.engine.utils import has_required_columns, partition_by_nullable
+
+if TYPE_CHECKING:
+    from rwa_calc.contracts.bundles import CounterpartyLookup
+
+logger = logging.getLogger(__name__)
+
+
+def propagate_facility_qrre_columns(
+    exposures: pl.LazyFrame,
+    facilities: pl.LazyFrame | None,
+) -> pl.LazyFrame:
+    """Coalesce QRRE-relevant facility-level columns onto every exposure row.
+
+    ``facility_undrawn`` exposures already carry ``is_revolving``,
+    ``is_qrre_transactor``, ``facility_limit`` (set in the facility-undrawn
+    select); loans / contingents have NULLs from the diagonal_relaxed concat.
+    This join fills them in from the parent facility.
+
+    Always returns a frame where ``is_revolving``, ``is_qrre_transactor``,
+    and ``facility_limit`` exist with safe defaults.
+    """
+    # Site B of the two-site QRRE coupling pinned by
+    # `_FACILITY_QRRE_COUPLED_COLUMNS` (stages/hierarchy/__init__.py). Site A
+    # lives in `facility_undrawn._undrawn_select_expressions` and projects the
+    # same column set when synthesising facility_undrawn rows; here we
+    # join+coalesce the parent facility's values onto loan / contingent
+    # exposure rows. The two sites use deliberately different shapes (project
+    # vs. join+coalesce) and must not be merged — only their column set is
+    # shared.
+    if facilities is not None:
+        exposures, qrre_schema = _join_facility_qrre_columns(exposures, facilities)
+    else:
+        qrre_schema = set(exposures.collect_schema().names())
+
+    # Ensure QRRE columns always exist with safe defaults.
+    # After the facility join branch above, these columns may or may not exist
+    # depending on the facility data. Reuse exp_schema from the join branch
+    # (or check fresh if we skipped the branch entirely).
+    exposures = _apply_qrre_defaults(exposures, qrre_schema)
+
+    # PRA PS1/26 Art. 121(4): the SCRA short-term window extends to self-
+    # liquidating trade-finance LCs. The flag lives on the facility row;
+    # drawn loans booked under a trade-LC facility have no facility_reference
+    # column to inherit from, so OR-aggregate the flag across the facilities
+    # of the same counterparty and broadcast it to every exposure of that
+    # counterparty.
+    # Coalesce preserves any explicit per-row value (e.g. on the synthetic
+    # facility_undrawn rows that already carry the flag from their source
+    # facility) and only fills nulls from the counterparty-level OR.
+    if facilities is not None:
+        exposures = _broadcast_trade_lc_flag(exposures, facilities)
+
+    return exposures
+
+
+@cites("CRR Art. 135")
+@cites("CRR Art. 136")
+@cites("CRR Art. 138")
+@cites("CRR Art. 139")
+def attach_counterparty_rating(
+    exposures: pl.LazyFrame,
+    counterparty_lookup: CounterpartyLookup,
+) -> pl.LazyFrame:
+    """Join counterparty rating fields onto every exposure row.
+
+    ``cqs`` and ``pd`` are used by SA / IRB calculators; ``internal_pd`` is
+    used by the classifier to gate IRB approach on internal-rating
+    availability; ``external_cqs`` is carried for audit trail; ``model_id``
+    (sourced from ``internal_model_id`` via the rating inheritance pipeline)
+    links to model_permissions for per-model approach gating.
+    """
+    cp_schema = set(counterparty_lookup.counterparties.collect_schema().names())
+    cp_select = [pl.col("counterparty_reference"), pl.col("cqs"), pl.col("pd")]
+    if "internal_pd" in cp_schema:
+        cp_select.append(pl.col("internal_pd"))
+    if "external_cqs" in cp_schema:
+        cp_select.append(pl.col("external_cqs"))
+    if "external_rating_is_issue_specific" in cp_schema:
+        cp_select.append(pl.col("external_rating_is_issue_specific"))
+    if "internal_model_id" in cp_schema:
+        cp_select.append(pl.col("internal_model_id"))
+
+    exposures = exposures.join(
+        counterparty_lookup.counterparties.select(cp_select),
+        on="counterparty_reference",
+        how="left",
+    )
+
+    # Ensure internal_pd, external_cqs, and model_id always exist for classifier
+    rating_defaults = []
+    if "internal_pd" not in cp_schema:
+        rating_defaults.append(pl.lit(None).cast(pl.Float64).alias("internal_pd"))
+    if "external_cqs" not in cp_schema:
+        rating_defaults.append(pl.lit(None).cast(pl.Int8).alias("external_cqs"))
+    # PRA PS1/26 Art. 139(2B): default provenance flag when no external rating
+    # resolved — treat as issue-specific (legacy behaviour, no disapplication).
+    if "external_rating_is_issue_specific" not in cp_schema:
+        rating_defaults.append(
+            pl.lit(True).cast(pl.Boolean).alias("external_rating_is_issue_specific")
+        )
+    if rating_defaults:
+        exposures = exposures.with_columns(rating_defaults)
+
+    # model_id: sourced from internal_model_id (rating inheritance pipeline).
+    # We know internal_model_id was joined from cp_schema above.
+    if "internal_model_id" in cp_schema:
+        return exposures.with_columns(pl.col("internal_model_id").alias("model_id")).drop(
+            "internal_model_id"
+        )
+    return exposures.with_columns(pl.lit(None).cast(pl.String).alias("model_id"))
+
+
+@cites("CRR Art. 131")
+@cites("CRR Art. 140")
+def apply_short_term_rating_override(
+    exposures: pl.LazyFrame,
+    ratings: pl.LazyFrame | None,
+) -> pl.LazyFrame:
+    """Apply per-exposure short-term rating override.
+
+    Short-term ECAI assessments under PRA PS1/26 Art. 120(2B) Table 4A and
+    Art. 122(3) Table 6A are issue-specific — each rating row attaches to a
+    single exposure via ``(scope_type, scope_id)``. When a short-term rating
+    row matches an exposure, its ``cqs`` overrides the counterparty-level
+    rating attached by ``attach_counterparty_rating`` and the derived
+    ``has_short_term_ecai`` flag is set to True, signalling the SA engine to
+    route via Table 4A / Table 6A.
+
+    Scope matching:
+
+    - ``scope_type='facility'``  -> matches the source facility's drawn loans,
+      its synthetic ``facility_undrawn`` row, and any descendant exposure via
+      ``parent_facility_reference`` / ``root_facility_reference``.
+    - ``scope_type='loan'``      -> matches the loan exposure with the same
+      ``exposure_reference`` and ``exposure_type='loan'``.
+    - ``scope_type='contingent'`` -> matches the contingent exposure with the
+      same ``exposure_reference`` and ``exposure_type='contingent'``.
+
+    Ties (multiple short-term ratings for the same exposure) are resolved by
+    picking the row with the lowest CQS, breaking ties by latest
+    ``rating_date``. This mirrors the external best-rating selection in
+    ``ratings.build_rating_inheritance_lazy``.
+
+    Always returns ``exposures`` augmented with a ``has_short_term_ecai``
+    boolean column (False when no override matched).
+    """
+    st_ratings = _prepare_short_term_lookup(ratings)
+    if st_ratings is None:
+        return exposures.with_columns(pl.lit(False).alias("has_short_term_ecai"))
+
+    exp_schema = set(exposures.collect_schema().names())
+    match_branches = _build_short_term_match_branches(exp_schema)
+
+    # Track which scope branches actually produce a match so we can
+    # coalesce the resulting cqs in priority order: loan > contingent >
+    # facility (most specific wins).
+    joined_scopes: list[str] = []
+    for scope, key_expr in match_branches:
+        scope_lookup = st_ratings.filter(pl.col("_st_scope_type") == scope).select(
+            [
+                pl.col("_st_cp"),
+                pl.col("_st_scope_id"),
+                pl.col("_st_cqs").alias(f"_st_{scope}_cqs"),
+            ]
+        )
+        exposures = exposures.with_columns(key_expr.alias(f"_match_key_{scope}"))
+        exposures = exposures.join(
+            scope_lookup,
+            left_on=["counterparty_reference", f"_match_key_{scope}"],
+            right_on=["_st_cp", "_st_scope_id"],
+            how="left",
+        ).drop(f"_match_key_{scope}")
+        joined_scopes.append(scope)
+
+    if not joined_scopes:
+        return exposures.with_columns(pl.lit(False).alias("has_short_term_ecai"))
+
+    # Coalesce in priority order: loan > contingent > facility (most
+    # specific scope wins).
+    priority = ["loan", "contingent", "facility"]
+    ordered = [s for s in priority if s in joined_scopes]
+    st_cqs_expr = pl.coalesce([pl.col(f"_st_{s}_cqs") for s in ordered])
+
+    # Override: when the short-term cqs is non-null, replace the cqs column
+    # and set has_short_term_ecai=True. SA Tables 4A / 6A are keyed off cqs
+    # only — rating_agency / rating_value are audit columns added later by
+    # the classifier and intentionally not overridden here.
+    has_st = st_cqs_expr.is_not_null()
+    exposures = exposures.with_columns(
+        [
+            has_st.alias("has_short_term_ecai"),
+            pl.when(has_st).then(st_cqs_expr).otherwise(pl.col("cqs")).cast(pl.Int8).alias("cqs"),
+        ]
+    )
+    return exposures.drop([f"_st_{s}_cqs" for s in joined_scopes])
+
+
+def enrich_with_property_coverage(
+    exposures: pl.LazyFrame,
+    collateral: pl.LazyFrame | None,
+) -> pl.LazyFrame:
+    """
+    Add property collateral coverage columns to exposures inline.
+
+    Calculates two separate values per exposure:
+    1. residential_collateral_value: For CRR Art. 123(c) retail threshold exclusion
+       (only residential property counts for threshold exclusion)
+    2. property_collateral_value: For retail_mortgage classification
+       (both residential AND commercial property qualify for mortgage treatment)
+
+    Uses .over() window functions for facility/counterparty allocation weights
+    instead of group_by + join-back, avoiding plan tree branching.
+
+    Args:
+        exposures: Unified exposures with hierarchy metadata
+        collateral: Collateral data with property_type and market_value
+
+    Returns:
+        Exposures with columns added: residential_collateral_value,
+        property_collateral_value, has_facility_property_collateral,
+        exposure_for_retail_threshold
+    """
+    # Per CRR Article 147, "total amount owed" = drawn amount only (not undrawn)
+    exposures = exposures.with_columns(
+        pl.col("drawn_amount").clip(lower_bound=0.0).alias("total_exposure_amount"),
+    )
+
+    required_cols = {
+        "beneficiary_reference",
+        "collateral_type",
+        "market_value",
+        "property_type",
+    }
+    if not has_required_columns(collateral, required_cols):
+        return exposures.with_columns(
+            [
+                pl.lit(0.0).alias("residential_collateral_value"),
+                pl.lit(0.0).alias("property_collateral_value"),
+                pl.lit(0.0).alias("residential_collateral_value_uncapped"),
+                pl.lit(0.0).alias("commercial_collateral_value_uncapped"),
+                pl.lit(False).alias("has_facility_property_collateral"),
+                pl.lit(False).alias("re_collateral_non_qualifying"),
+                pl.col("total_exposure_amount").alias("exposure_for_retail_threshold"),
+            ]
+        )
+
+    # Single filter for all property collateral; split residential inline
+    all_property_collateral = collateral.filter(
+        pl.col("collateral_type").str.to_lowercase() == "real_estate"
+    )
+    # PRA PS1/26 Art. 124(4): a single non-qualifying RE component (Art. 124A
+    # failure, e.g. valuation-independence breach) forces the WHOLE mixed-RE
+    # exposure to Art. 124J. Track per-beneficiary whether any RE collateral
+    # row fails the qualifying test so the classifier can fire the gate.
+    # ``is_qualifying_re`` has no schema default — null means "unreported"
+    # and is treated as qualifying here (fill_null(True) is a value fill).
+    is_non_qualifying_re = pl.col("is_qualifying_re").fill_null(True) == False  # noqa: E712
+
+    # Multi-level linking with .over() allocation weights
+    exposures = _join_property_collateral_multi_level(
+        exposures,
+        all_property_collateral,
+        is_non_qualifying_re=is_non_qualifying_re,
+    )
+
+    # Fill nulls, then cap at exposure amount and derive threshold.
+    # Preserve the UNCAPPED residential / commercial RE collateral values
+    # for the loan-splitter: the PRA PS1/26 Art. 124(4) pro-rata split is by
+    # raw collateral value (and the 0.55xV cap is also on raw property value),
+    # so the per-exposure cap applied below — which exists for the CRR retail
+    # threshold — must not distort the split shares.
+    exposures = (
+        exposures.with_columns(
+            [
+                pl.col("residential_collateral_value").fill_null(0.0),
+                pl.col("property_collateral_value").fill_null(0.0),
+                pl.col("re_collateral_non_qualifying").fill_null(False),
+            ]
+        )
+        .with_columns(
+            [
+                pl.col("residential_collateral_value").alias(
+                    "residential_collateral_value_uncapped"
+                ),
+                (pl.col("property_collateral_value") - pl.col("residential_collateral_value"))
+                .clip(lower_bound=0.0)
+                .alias("commercial_collateral_value_uncapped"),
+            ]
+        )
+        .with_columns(
+            [
+                pl.min_horizontal("residential_collateral_value", "total_exposure_amount").alias(
+                    "residential_collateral_value"
+                ),
+                pl.min_horizontal("property_collateral_value", "total_exposure_amount").alias(
+                    "property_collateral_value"
+                ),
+                (
+                    pl.col("total_exposure_amount")
+                    - pl.min_horizontal("residential_collateral_value", "total_exposure_amount")
+                ).alias("exposure_for_retail_threshold"),
+            ]
+        )
+    )
+
+    return exposures
+
+
+def enrich_with_lending_group(
+    exposures: pl.LazyFrame,
+    lending_mappings: pl.LazyFrame | None,
+) -> pl.LazyFrame:
+    """
+    Add lending group reference and exposure totals to each exposure.
+
+    Uses .over() window functions to compute group totals inline instead of
+    group_by + join-back, avoiding plan tree branching.
+
+    Per CRR Art. 123(c), the adjusted_exposure (excluding residential property)
+    is used for retail threshold testing. When a counterparty is not part of
+    an explicit lending group, it is treated as a group-of-one per CRR Art.
+    4(1)(39) ("group of connected clients") — totals are aggregated across
+    the counterparty's own exposures rather than leaving the per-row value.
+
+    Args:
+        exposures: Exposures with property coverage columns already added
+        lending_mappings: Lending group parent-child mappings, or None when
+            no lending-group table was supplied — every counterparty is
+            then a group-of-one (identical to an empty mappings table)
+
+    Returns:
+        Exposures with lending_group_reference, lending_group_total_exposure,
+        and lending_group_adjusted_exposure columns added
+    """
+    if lending_mappings is None:
+        # No lending-group table: every row's group reference is null, so
+        # the null-partition fallback below aggregates per counterparty —
+        # exactly the empty-mappings behaviour.
+        exposures = exposures.with_columns(
+            pl.lit(None).cast(pl.String).alias("lending_group_reference")
+        )
+    else:
+        # Build lending group membership
+        lending_groups = lending_mappings.select(
+            [
+                pl.col("parent_counterparty_reference").alias("lending_group_reference"),
+                pl.col("child_counterparty_reference").alias("member_counterparty_reference"),
+            ]
+        )
+
+        parent_as_member = lending_mappings.select(
+            [
+                pl.col("parent_counterparty_reference").alias("lending_group_reference"),
+                pl.col("parent_counterparty_reference").alias("member_counterparty_reference"),
+            ]
+        ).unique()
+
+        all_members = pl.concat(
+            [lending_groups, parent_as_member],
+            how="vertical",
+        ).unique(subset=["member_counterparty_reference"], keep="first")
+
+        # Join to get lending group reference
+        exposures = exposures.join(
+            all_members,
+            left_on="counterparty_reference",
+            right_on="member_counterparty_reference",
+            how="left",
+        )
+
+    # .over() window functions for group totals (no self-join!).
+    # When no lending group, aggregate over counterparty_reference so the
+    # retail threshold test sees the obligor's full exposure, not a single
+    # line. lending_group_reference is left-join nullable, so the null-
+    # partition guard prevents pooling unrelated unmapped rows.
+    exposures = exposures.with_columns(
+        [
+            partition_by_nullable(
+                pl.col("drawn_amount").clip(lower_bound=0.0).sum().over("lending_group_reference")
+                + pl.col("nominal_amount").sum().over("lending_group_reference"),
+                "lending_group_reference",
+                pl.col("drawn_amount").clip(lower_bound=0.0).sum().over("counterparty_reference")
+                + pl.col("nominal_amount").sum().over("counterparty_reference"),
+            ).alias("lending_group_total_exposure"),
+            partition_by_nullable(
+                pl.col("exposure_for_retail_threshold").sum().over("lending_group_reference"),
+                "lending_group_reference",
+                pl.col("exposure_for_retail_threshold").sum().over("counterparty_reference"),
+            ).alias("lending_group_adjusted_exposure"),
+        ]
+    )
+
+    return exposures
+
+
+def add_collateral_ltv(
+    exposures: pl.LazyFrame,
+    collateral: pl.LazyFrame | None,
+) -> pl.LazyFrame:
+    """
+    Add LTV and property metadata from collateral for real estate risk weights.
+
+    Joins collateral property_ltv, property_type, is_income_producing, and
+    is_qualifying_re to exposures where collateral is linked via
+    beneficiary_reference. For mortgages and commercial RE, LTV determines risk
+    weight. For CRE (CRR Art. 126), income cover status determines 50% vs 100%
+    risk weight. Non-qualifying RE (Art. 124J) gets separate treatment under B31.
+
+    Supports three levels of collateral linking based on beneficiary_type:
+    1. Direct (exposure/loan): beneficiary_reference matches exposure_reference
+    2. Facility: beneficiary_reference matches parent_facility_reference
+    3. Counterparty: beneficiary_reference matches counterparty_reference
+
+    Args:
+        exposures: Unified exposures with exposure_reference
+        collateral: Collateral data with beneficiary_reference and property_ltv (optional)
+
+    Returns:
+        Exposures with ltv, property_type, has_income_cover, and
+        is_qualifying_re columns added
+    """
+    # Check if collateral is valid for LTV processing
+    # Requires beneficiary_reference and property_ltv columns
+    required_cols = {"beneficiary_reference", "property_ltv"}
+    if not has_required_columns(collateral, required_cols):
+        return _add_ltv_defaults_for_missing_collateral(exposures)
+
+    # Filter for collateral with LTV data
+    ltv_collateral = collateral.filter(pl.col("property_ltv").is_not_null())
+
+    # Multi-level linking: separate collateral by beneficiary_type, then
+    # coalesce direct -> facility -> counterparty so the most specific
+    # collateral wins. The collateral frame is sealed at the loader edge,
+    # so the four property columns always exist; ``is_income_producing``
+    # is loader-defaulted (schema default False), so no null fill needed.
+    common_cols = (
+        pl.col("property_type"),
+        pl.col("is_income_producing").alias("has_income_cover"),
+        pl.col("is_qualifying_re"),
+        pl.col("prior_charge_ltv"),
+    )
+    direct_ltv = _level_ltv_lookup(
+        ltv_collateral,
+        filter_expr=pl.col("beneficiary_type").str.to_lowercase().is_in(["exposure", "loan"]),
+        prefix="direct",
+        common_cols=common_cols,
+    )
+    facility_ltv = _level_ltv_lookup(
+        ltv_collateral,
+        filter_expr=pl.col("beneficiary_type").str.to_lowercase() == "facility",
+        prefix="facility",
+        common_cols=common_cols,
+    )
+    counterparty_ltv = _level_ltv_lookup(
+        ltv_collateral,
+        filter_expr=pl.col("beneficiary_type").str.to_lowercase() == "counterparty",
+        prefix="cp",
+        common_cols=common_cols,
+    )
+
+    # Join all three levels onto the exposures frame.
+    exposures = (
+        exposures.join(
+            direct_ltv,
+            left_on="exposure_reference",
+            right_on="direct_ref",
+            how="left",
+        )
+        .join(
+            facility_ltv,
+            left_on="parent_facility_reference",
+            right_on="facility_ref",
+            how="left",
+        )
+        .join(
+            counterparty_ltv,
+            left_on="counterparty_reference",
+            right_on="cp_ref",
+            how="left",
+        )
+    )
+
+    return _coalesce_ltv_levels(exposures, prefixes=("direct", "facility", "cp"))
+
+
+def _join_facility_qrre_columns(
+    exposures: pl.LazyFrame,
+    facilities: pl.LazyFrame,
+) -> tuple[pl.LazyFrame, set[str]]:
+    """Join facility-side QRRE / limit / termination columns onto exposures.
+
+    Scratch: facility-side QRRE / limit / termination columns join as
+    ``_fac_*``, get coalesced into their unprefixed exposure-level
+    counterparts (``is_revolving``, ``is_qrre_transactor``, ``facility_limit``,
+    ``facility_termination_date``), then dropped via the scratch aliases.
+    The facilities frame is sealed at the loader edge, so every source
+    column exists and the Boolean ones are non-null.
+    """
+    # (facility column, scratch alias, exposure-level column, fill-null bool default)
+    fac_specs: tuple[tuple[str, str, str, bool], ...] = (
+        ("is_revolving", "_fac_revolving", "is_revolving", True),
+        ("is_qrre_transactor", "_fac_transactor", "is_qrre_transactor", True),
+        ("limit", "_fac_limit", "facility_limit", False),
+        (
+            "facility_termination_date",
+            "_fac_termination_date",
+            "facility_termination_date",
+            False,
+        ),
+    )
+
+    fac_select: list[pl.Expr] = [pl.col("facility_reference").alias("_fac_ref")]
+    fac_select.extend(pl.col(src_col).alias(alias) for src_col, alias, _, _ in fac_specs)
+
+    exposures = exposures.join(
+        facilities.select(fac_select),
+        left_on="parent_facility_reference",
+        right_on="_fac_ref",
+        how="left",
+    )
+
+    # Single schema check covers both QRRE coalesce and default columns
+    exp_schema = set(exposures.collect_schema().names())
+    exposures = exposures.with_columns(
+        [
+            _build_qrre_coalesce_expr(alias, exp_col, exp_schema, fill_false)
+            for _src_col, alias, exp_col, fill_false in fac_specs
+        ]
+    )
+
+    # Drop temporary join columns
+    exposures = exposures.drop([alias for _src_col, alias, _exp_col, _fill in fac_specs])
+    return exposures, exp_schema
+
+
+def _join_property_collateral_multi_level(
+    exposures: pl.LazyFrame,
+    all_property_collateral: pl.LazyFrame,
+    *,
+    is_non_qualifying_re: pl.Expr,
+) -> pl.LazyFrame:
+    """
+    Join property collateral at direct/facility/counterparty levels.
+
+    Uses a single conditional group_by across all property collateral and
+    3 joins (one per level) instead of 6 separate aggregations + 6 joins.
+    Allocation weights use .over() window functions.
+
+    Args:
+        exposures: Exposures with total_exposure_amount column
+        all_property_collateral: All property collateral (residential + commercial);
+            residential rows are filtered inline via ``filter(is_residential)``
+            within the conditional ``group_by`` aggregate.
+        is_non_qualifying_re: Per-collateral-row predicate flagging RE that
+            fails Art. 124A (``is_qualifying_re == False``); aggregated to a
+            per-beneficiary ``re_collateral_non_qualifying`` flag for the
+            PRA PS1/26 Art. 124(4) all-or-nothing gate.
+
+    Returns:
+        Exposures with residential_collateral_value, property_collateral_value,
+        re_collateral_non_qualifying, and has_facility_property_collateral
+        columns added
+    """
+    bt_lower = pl.col("beneficiary_type").str.to_lowercase()
+    is_residential = pl.col("property_type").str.to_lowercase() == "residential"
+
+    # Scratch: a single conditional group_by produces `_level` (direct /
+    # facility / counterparty), `_res` (residential market value sum), and
+    # `_prop` (all-property market value sum). The aggregate is then split
+    # by `_level` into three frames; per-level columns are renamed with
+    # suffixes (`_res_d`/`_prop_d`, `_res_f`/`_prop_f`, `_res_c`/`_prop_c`)
+    # so the three joins below can carry their own values without collision.
+    # All scratch columns are coalesced and dropped before the helper returns.
+    # Single conditional group_by: 6 aggregates in one pass
+    coll_agg = (
+        all_property_collateral.with_columns(
+            pl.when(bt_lower.is_in(["exposure", "loan"]))
+            .then(pl.lit("direct"))
+            .when(bt_lower == "facility")
+            .then(pl.lit("facility"))
+            .when(bt_lower == "counterparty")
+            .then(pl.lit("counterparty"))
+            .otherwise(pl.lit("direct"))
+            .alias("_level"),
+        )
+        .group_by(["_level", "beneficiary_reference"])
+        .agg(
+            [
+                pl.col("market_value").filter(is_residential).sum().alias("_res"),
+                pl.col("market_value").sum().alias("_prop"),
+                is_non_qualifying_re.any().alias("_nonqual"),
+            ]
+        )
+    )
+
+    # Split and rename for per-level joins
+    coll_direct = (
+        coll_agg.filter(pl.col("_level") == "direct")
+        .drop("_level")
+        .rename({"_res": "_res_d", "_prop": "_prop_d", "_nonqual": "_nonqual_d"})
+    )
+    coll_facility = (
+        coll_agg.filter(pl.col("_level") == "facility")
+        .drop("_level")
+        .rename({"_res": "_res_f", "_prop": "_prop_f", "_nonqual": "_nonqual_f"})
+    )
+    coll_cp = (
+        coll_agg.filter(pl.col("_level") == "counterparty")
+        .drop("_level")
+        .rename({"_res": "_res_c", "_prop": "_prop_c", "_nonqual": "_nonqual_c"})
+    )
+
+    # .over() window functions for allocation weights (no self-join!).
+    # Both partition keys are nullable in this frame: direct exposures
+    # (no facility) have null parent_facility_reference; null
+    # counterparty_reference can arise from upstream join misses. The
+    # null-partition guard prevents pooling unrelated rows.
+    exposures = exposures.with_columns(
+        [
+            partition_by_nullable(
+                pl.col("total_exposure_amount").sum().over("parent_facility_reference"),
+                "parent_facility_reference",
+                pl.col("total_exposure_amount"),
+            ).alias("facility_total"),
+            partition_by_nullable(
+                pl.col("total_exposure_amount").sum().over("counterparty_reference"),
+                "counterparty_reference",
+                pl.col("total_exposure_amount"),
+            ).alias("cp_total"),
+        ]
+    )
+
+    # 3 joins (one per level) instead of 6
+    exposures = (
+        exposures.join(
+            coll_direct,
+            left_on="exposure_reference",
+            right_on="beneficiary_reference",
+            how="left",
+        )
+        .join(
+            coll_facility,
+            left_on="parent_facility_reference",
+            right_on="beneficiary_reference",
+            how="left",
+        )
+        .join(
+            coll_cp,
+            left_on="counterparty_reference",
+            right_on="beneficiary_reference",
+            how="left",
+        )
+    )
+
+    # Pro-rata weights + combine all levels in one batch
+    exposures = exposures.with_columns(
+        [
+            pl.when(pl.col("facility_total") > 0)
+            .then(pl.col("total_exposure_amount") / pl.col("facility_total"))
+            .otherwise(pl.lit(0.0))
+            .alias("facility_weight"),
+            pl.when(pl.col("cp_total") > 0)
+            .then(pl.col("total_exposure_amount") / pl.col("cp_total"))
+            .otherwise(pl.lit(0.0))
+            .alias("cp_weight"),
+        ]
+    )
+
+    exposures = exposures.with_columns(
+        [
+            (
+                pl.col("_res_d").fill_null(0.0)
+                + (pl.col("_res_f").fill_null(0.0) * pl.col("facility_weight"))
+                + (pl.col("_res_c").fill_null(0.0) * pl.col("cp_weight"))
+            ).alias("residential_collateral_value"),
+            (
+                pl.col("_prop_d").fill_null(0.0)
+                + (pl.col("_prop_f").fill_null(0.0) * pl.col("facility_weight"))
+                + (pl.col("_prop_c").fill_null(0.0) * pl.col("cp_weight"))
+            ).alias("property_collateral_value"),
+            (
+                (pl.col("_prop_d").fill_null(0.0) > 0)
+                | (pl.col("_prop_f").fill_null(0.0) > 0)
+                | (pl.col("_prop_c").fill_null(0.0) > 0)
+            ).alias("has_facility_property_collateral"),
+            (
+                pl.col("_nonqual_d").fill_null(False)
+                | pl.col("_nonqual_f").fill_null(False)
+                | pl.col("_nonqual_c").fill_null(False)
+            ).alias("re_collateral_non_qualifying"),
+        ]
+    )
+
+    # Drop intermediate columns
+    return exposures.drop(
+        [
+            "_res_d",
+            "_res_f",
+            "_res_c",
+            "_prop_d",
+            "_prop_f",
+            "_prop_c",
+            "_nonqual_d",
+            "_nonqual_f",
+            "_nonqual_c",
+            "facility_total",
+            "cp_total",
+            "facility_weight",
+            "cp_weight",
+        ]
+    )
+
+
+def _level_ltv_lookup(
+    ltv_collateral: pl.LazyFrame,
+    *,
+    filter_expr: pl.Expr,
+    prefix: str,
+    common_cols: tuple[pl.Expr, pl.Expr, pl.Expr, pl.Expr],
+) -> pl.LazyFrame:
+    """Build a single-level LTV lookup frame with prefixed columns.
+
+    Filters ``ltv_collateral`` to a single beneficiary level, then projects
+    ``beneficiary_reference`` and ``property_ltv`` plus the four optional
+    property columns (property_type, income_cover, qualifying_re,
+    prior_charge_ltv) all aliased with ``{prefix}_``. Deduplicated on the
+    prefixed reference so a beneficiary appearing twice in collateral does
+    not produce duplicate exposure rows after the join.
+    """
+    prop_type, income_cover, qualifying_re, prior_charge_ltv = common_cols
+    return (
+        ltv_collateral.filter(filter_expr)
+        .select(
+            [
+                pl.col("beneficiary_reference").alias(f"{prefix}_ref"),
+                pl.col("property_ltv").alias(f"{prefix}_ltv"),
+                prop_type.alias(f"{prefix}_property_type"),
+                income_cover.alias(f"{prefix}_income_cover"),
+                qualifying_re.alias(f"{prefix}_qualifying_re"),
+                prior_charge_ltv.alias(f"{prefix}_prior_charge_ltv"),
+            ]
+        )
+        .unique(subset=[f"{prefix}_ref"], keep="first")
+    )
+
+
+def _coalesce_ltv_levels(
+    exposures: pl.LazyFrame,
+    *,
+    prefixes: tuple[str, ...],
+) -> pl.LazyFrame:
+    """Collapse per-level LTV columns onto the unified exposure frame.
+
+    For each output column (ltv, property_type, has_income_cover,
+    is_qualifying_re, prior_charge_ltv), coalesce in declared ``prefixes``
+    order — earliest non-null wins. ``has_income_cover`` additionally
+    defaults to ``False`` when no level provided a value. All scratch
+    ``{prefix}_*`` columns are dropped at the end.
+    """
+    # (source_suffix, output_col, fill_null_default_or_sentinel)
+    _NO_DEFAULT: object = object()
+    coalesce_specs: list[tuple[str, str, object]] = [
+        ("ltv", "ltv", _NO_DEFAULT),
+        ("property_type", "property_type", _NO_DEFAULT),
+        ("income_cover", "has_income_cover", False),
+        ("qualifying_re", "is_qualifying_re", _NO_DEFAULT),
+        ("prior_charge_ltv", "prior_charge_ltv", _NO_DEFAULT),
+    ]
+
+    coalesces: list[pl.Expr] = []
+    drop_cols: list[str] = []
+    for source_suffix, output_col, default in coalesce_specs:
+        expr = pl.coalesce(*[pl.col(f"{p}_{source_suffix}") for p in prefixes])
+        if default is not _NO_DEFAULT:
+            expr = expr.fill_null(default)
+        coalesces.append(expr.alias(output_col))
+        drop_cols.extend(f"{p}_{source_suffix}" for p in prefixes)
+
+    return exposures.with_columns(coalesces).drop(drop_cols)
+
+
+def _add_ltv_defaults_for_missing_collateral(exposures: pl.LazyFrame) -> pl.LazyFrame:
+    """Append null/default LTV columns the caller has not already populated.
+
+    Loan / contingent fixtures may carry exposure-level ``ltv`` /
+    ``property_type`` / ``has_income_cover`` / ``is_qualifying_re`` /
+    ``prior_charge_ltv`` values (e.g. CRE Art. 126(2)(d) scenarios where the
+    LTV and income-cover flags live on the loan rather than a collateral row);
+    overwriting them here would silently break the SA real-estate branch
+    downstream.
+    """
+    existing = set(exposures.collect_schema().names())
+    # (output column, default expression)
+    default_specs: tuple[tuple[str, pl.Expr], ...] = (
+        ("ltv", pl.lit(None).cast(pl.Float64).alias("ltv")),
+        ("property_type", pl.lit(None).cast(pl.Utf8).alias("property_type")),
+        ("has_income_cover", pl.lit(False).alias("has_income_cover")),
+        ("is_qualifying_re", pl.lit(None).cast(pl.Boolean).alias("is_qualifying_re")),
+        ("prior_charge_ltv", pl.lit(None).cast(pl.Float64).alias("prior_charge_ltv")),
+    )
+    defaults = [expr for col, expr in default_specs if col not in existing]
+    return exposures.with_columns(defaults) if defaults else exposures
+
+
+def _prepare_short_term_lookup(ratings: pl.LazyFrame | None) -> pl.LazyFrame | None:
+    """Filter, sort and materialise the short-term rating lookup.
+
+    Returns ``None`` if no short-term rows are available (i.e. ``ratings`` is
+    ``None`` or the filtered set is empty). The caller treats ``None`` as "no
+    override applies — set ``has_short_term_ecai=False``".
+    """
+    if ratings is None:
+        return None
+
+    # Filter to candidate short-term rows. Drop rows missing the required
+    # scope tuple — loader-side DQ flags those as DQ-RT-ST1 / DQ-RT-ST2
+    # errors; here we silently ignore them so the pipeline keeps running.
+    # ``is_short_term`` is loader-defaulted to False (Boolean schema default),
+    # so no null fill is needed.
+    st_ratings = ratings.filter(
+        pl.col("is_short_term")
+        & pl.col("scope_type").is_not_null()
+        & pl.col("scope_id").is_not_null()
+        & pl.col("counterparty_reference").is_not_null()
+    ).select(
+        [
+            pl.col("counterparty_reference").alias("_st_cp"),
+            pl.col("scope_type").alias("_st_scope_type"),
+            pl.col("scope_id").alias("_st_scope_id"),
+            pl.col("cqs").alias("_st_cqs"),
+            pl.col("rating_date").alias("_st_rating_date"),
+        ]
+    )
+
+    # Per-scope best-rating selection: lowest CQS, then latest date.
+    st_ratings = (
+        st_ratings.sort(
+            ["_st_cqs", "_st_rating_date"],
+            descending=[False, True],
+            nulls_last=True,
+        )
+        .group_by(["_st_cp", "_st_scope_type", "_st_scope_id"])
+        .first()
+    )
+
+    # Materialise the small short-term lookup eagerly so the three scope-
+    # specific joins below can re-use it without re-evaluating the sort.
+    st_ratings_df = st_ratings.collect()
+    if st_ratings_df.height == 0:
+        return None
+    return st_ratings_df.lazy()
+
+
+def _build_short_term_match_branches(exp_schema: set[str]) -> list[tuple[str, pl.Expr]]:
+    """Build ``(scope_type, key_expression)`` pairs for each available scope.
+
+    An exposure can satisfy multiple scope_types simultaneously (e.g. a loan
+    exposure also inherits its parent facility's short-term rating); the
+    caller left-joins each branch and coalesces in priority order so the most
+    specific scope wins (loan > contingent > facility).
+    """
+    match_branches: list[tuple[str, pl.Expr]] = []
+    # facility scope: any exposure whose parent or root facility id matches
+    has_parent = "parent_facility_reference" in exp_schema
+    has_root = "root_facility_reference" in exp_schema
+    if has_parent or has_root:
+        facility_key_expr = pl.coalesce(
+            [
+                pl.col("parent_facility_reference")
+                if has_parent
+                else pl.lit(None, dtype=pl.String),
+                pl.col("root_facility_reference") if has_root else pl.lit(None, dtype=pl.String),
+            ]
+        )
+        match_branches.append(("facility", facility_key_expr))
+    # loan / contingent scope: match by exposure_reference + exposure_type
+    if "exposure_type" in exp_schema and "exposure_reference" in exp_schema:
+        match_branches.append(
+            (
+                "loan",
+                pl.when(pl.col("exposure_type") == "loan")
+                .then(pl.col("exposure_reference"))
+                .otherwise(pl.lit(None, dtype=pl.String)),
+            )
+        )
+        match_branches.append(
+            (
+                "contingent",
+                pl.when(pl.col("exposure_type") == "contingent")
+                .then(pl.col("exposure_reference"))
+                .otherwise(pl.lit(None, dtype=pl.String)),
+            )
+        )
+    return match_branches
+
+
+def _build_qrre_coalesce_expr(
+    alias: str,
+    exp_col: str,
+    exp_schema: set[str],
+    fill_false: bool,
+) -> pl.Expr:
+    """Build the coalesce / fallback expression for one QRRE column.
+
+    When ``exp_col`` already exists on the exposures frame, coalesce its value
+    with the joined ``alias`` scratch column; otherwise project the scratch
+    column directly. ``fill_false=True`` applies ``fill_null(False)`` so the
+    output column is non-null boolean.
+    """
+    expr = pl.coalesce(pl.col(exp_col), pl.col(alias)) if exp_col in exp_schema else pl.col(alias)
+    if fill_false:
+        expr = expr.fill_null(False)
+    return expr.alias(exp_col)
+
+
+def _apply_qrre_defaults(exposures: pl.LazyFrame, qrre_schema: set[str]) -> pl.LazyFrame:
+    """Ensure ``is_revolving``, ``is_qrre_transactor`` and ``facility_limit`` exist."""
+    default_specs: tuple[tuple[str, pl.Expr], ...] = (
+        ("is_revolving", pl.lit(False).alias("is_revolving")),
+        ("is_qrre_transactor", pl.lit(False).alias("is_qrre_transactor")),
+        ("facility_limit", pl.lit(None).cast(pl.Float64).alias("facility_limit")),
+    )
+    default_cols = [expr for col, expr in default_specs if col not in qrre_schema]
+    if default_cols:
+        exposures = exposures.with_columns(default_cols)
+    return exposures
+
+
+def _broadcast_trade_lc_flag(
+    exposures: pl.LazyFrame,
+    facilities: pl.LazyFrame,
+) -> pl.LazyFrame:
+    """OR-aggregate ``is_short_term_trade_lc`` per counterparty and broadcast.
+
+    Coalesces with any explicit per-row value already on the exposures frame
+    (e.g. synthetic facility_undrawn rows carrying the flag from their source
+    facility) and only fills nulls from the counterparty-level OR.
+    """
+    cp_trade_lc = facilities.group_by("counterparty_reference").agg(
+        pl.col("is_short_term_trade_lc").any().alias("_cp_trade_lc")
+    )
+    exposures = exposures.join(
+        cp_trade_lc,
+        on="counterparty_reference",
+        how="left",
+    )
+    if "is_short_term_trade_lc" in exposures.collect_schema().names():
+        exposures = exposures.with_columns(
+            pl.coalesce(
+                pl.col("is_short_term_trade_lc"),
+                pl.col("_cp_trade_lc"),
+            )
+            .fill_null(False)
+            .alias("is_short_term_trade_lc")
+        )
+    else:
+        exposures = exposures.with_columns(
+            pl.col("_cp_trade_lc").fill_null(False).alias("is_short_term_trade_lc")
+        )
+    return exposures.drop("_cp_trade_lc")
