+++ src/rwa_calc/engine/stages/hierarchy/stage.py
