+++ tests/contracts/test_no_raw_over_on_nullable_keys.py
@@ -46,13 +46,13 @@ ALLOWLIST: frozenset[str] = frozenset(
 # justification comment naming the filter or invariant that makes the key
 # non-null at this site.
 LINE_ALLOWLIST: dict[str, frozenset[int]] = {
-    # _build_rating_inheritance_lazy: the per_agency_latest filter adds
+    # build_rating_inheritance_lazy: the per_agency_latest filter adds
     # `counterparty_reference.is_not_null()` to the upstream filter, so these
     # `.over("counterparty_reference")` calls operate on a frame with no
     # null-keyed rows. See the comment block above the per_agency_latest
-    # filter in engine/hierarchy.py. (Line numbers shifted +13 by the Phase 3
-    # CounterpartyLookup producer-seal block added above them.)
-    "src/rwa_calc/engine/hierarchy.py": frozenset({450, 451}),
+    # filter in engine/stages/hierarchy/ratings.py. (Re-keyed from
+    # engine/hierarchy.py:450/451 by the Phase 4 Slice 2 package split.)
+    "src/rwa_calc/engine/stages/hierarchy/ratings.py": frozenset({126, 127}),
 }
